Split gui tabs into files, moved constants to cache
