+++ modis/__init__.py
@@ -94,12 +94,12 @@ def log_init():
     logger = logging.getLogger(__name__)
 
     # Set log level
-    data = datatools.get()
-    if "log_level" in data:
-        logger.setLevel(data["log_level"])
+    datatools.get()
+    if "log_level" in datatools.data:
+        logger.setLevel(datatools.data["log_level"])
     else:
-        data["log_level"] = "INFO"
-        datatools.write(data)
+        datatools.data["log_level"] = "INFO"
+        datatools.write()
         logger.setLevel("INFO")
 
     # Setup logging format
